+++ src/components/donation/Steps/DonateMethods/Stripe/useRhf.ts
@@ -1,0 +1,77 @@
+import { yupResolver } from "@hookform/resolvers/yup";
+import { useController, useForm } from "react-hook-form";
+import { schema, stringNumber } from "schemas/shape";
+import { requiredString } from "schemas/string";
+import type { Currency } from "types/components";
+import { DEFAULT_PROGRAM, usdOption } from "../../common/constants";
+import type { FormValues as FV, FormProps, OnIncrement } from "./types";
+
+export function useRhf(props: Omit<FormProps, "currencies">) {
+  const initial: FV = {
+    amount: "",
+    currency: props.defaultCurr || usdOption,
+    frequency: "subscription",
+    program: DEFAULT_PROGRAM,
+  };
+
+  const currencyKey: keyof FV = "currency";
+  const {
+    control,
+    handleSubmit,
+    register,
+    setValue,
+    getValues,
+    trigger,
+    formState: { errors },
+  } = useForm<FV>({
+    defaultValues: props.details || initial,
+    resolver: yupResolver(
+      schema<FV>({
+        frequency: requiredString,
+        amount: stringNumber(
+          (s) => s.required("Please enter an amount"),
+          (n) =>
+            n
+              .positive("Amount must be greater than 0")
+              .when(currencyKey, (values, schema) => {
+                const [currency] = values as [Currency | undefined];
+                return currency?.min
+                  ? schema.min(currency.min, "less than min")
+                  : schema;
+              })
+        ),
+      })
+    ),
+  });
+
+  const { field: frequency } = useController<FV, "frequency">({
+    control: control,
+    name: "frequency",
+  });
+
+  const { field: currency } = useController<FV, "currency">({
+    control: control,
+    name: "currency",
+  });
+
+  const { field: program } = useController<FV, "program">({
+    control: control,
+    name: "program",
+  });
+
+  const onIncrement: OnIncrement = (inc) => {
+    const amntNum = Number(getValues("amount"));
+    if (Number.isNaN(amntNum)) return trigger("amount");
+    setValue("amount", `${inc + amntNum}`);
+  };
+
+  return {
+    frequency,
+    currency,
+    program,
+    onIncrement,
+    register,
+    handleSubmit,
+    errors,
+  };
+}
